Address review comments: - Rename KeyProviderForTesting.java to MockAesKeyProvider.java - Added class level comments for copied / moved code. - Break TestBoundedRegionGroupingStrategy to TestBoundedRegionGroupingStrategyWPETool refactoring tests needing WALPerformanceEvaluation into it
